ds-nn-m4/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
boto3==1.35.80
botocore==1.35.80
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
jmespath==1.0.1
lxml==5.3.0
numpy==2.2.0
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1